overlay: SimpleDeviceConfig: Compact cached app heaps in the…
device_config: Compact cached app heaps in the background

Background ART heap compaction for cached apps has been recommended for
Android Go devices for a while, and Google has enabled it for the Pixel
5 on RQ1A.201205.010. This feature should help fit more cached apps in
memory.

device_config: Allow more cached apps in the background

Google has provisioned this on the Pixel 5 (RQ1A.201205.010) to utilize
the 8 GB of RAM better, and I have also observed that a lightweight
selection of apps was hitting the cached app limit before exhausting RAM
on another device with 8 GB of RAM.

This shouldn't cause adverse effects on devices with less RAM (e.g. 4
GB) as the low memory killer should kick in long before this limit on
such devices.

Change-Id: I59410216640beea958841ebcd6d2277e6c734091